Move time limit edit form to separate file
